+++ js/data.js
@@ -1,0 +1,74 @@
+import { getRandomArrayElement, getRandomInteger, generateRandomId } from './util.js';
+
+const PICTURE_COUNT = 25;
+const LIKE_MIN_COUNT = 15;
+const LIKE_MAX_COUNT = 200;
+const COMMENT_MIN_COUNT = 0;
+const COMMENT_MAX_COUNT = 30;
+const AVATAR_MIN_COUNT = 1;
+const AVATAR_MAX_COUNT = 6;
+const MESSAGE_MIN_COUNT = 1;
+const MESSAGE_MAX_COUNT = 2;
+
+const MESSAGE_TEXT = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+const DESCRIPTIONS = [
+  'Тег #закат, #друзья, #природа, #красота',
+  'Мода пройдет, стиль останется. #красота',
+  'Воскресное селфи.',
+  'Даже мой кофе хочет кофе.',
+  'Мы с котом. Нам не нужен никто.',
+  'Драйв и кайф!',
+  'История, рассказанная через объектив.',
+  'Море — лучший лекарь для усталой души',
+  'Настоящий друг с тобой, когда ты неправ. Когда ты прав, всякий будет с тобой',
+  'Идеальное место для отдыха и вдохновения. #природа #горы #озёра #отдых #вдохновение',
+];
+const NAMES = [
+  'Игорь',
+  'Артём',
+  'Екатерина',
+  'Ира',
+  'Олег',
+  'Катя',
+  'Светлана',
+  'Димон',
+  'Ксюша',
+  'Лампа'
+];
+
+const createMessage = () => Array.from(
+  {length: getRandomInteger(MESSAGE_MIN_COUNT, MESSAGE_MAX_COUNT)},
+  () => getRandomArrayElement(MESSAGE_TEXT),
+).join(' ');
+
+const createComment = () => ({
+  id: generateRandomId(),
+  avatar: `img/avatar-${getRandomInteger(AVATAR_MIN_COUNT, AVATAR_MAX_COUNT)}.svg`,
+  message: createMessage(),
+  name: getRandomArrayElement(NAMES),
+});
+
+const createPicture = (index) => ({
+  id: index,
+  url: `photos/${index}.jpg`,
+  description: getRandomArrayElement(DESCRIPTIONS),
+  likes: getRandomInteger(LIKE_MIN_COUNT, LIKE_MAX_COUNT),
+  comments: Array.from(
+    {length: getRandomInteger(COMMENT_MIN_COUNT, COMMENT_MAX_COUNT)},
+    createComment
+  )
+});
+
+const getPictures = () => Array.from(
+  {length: PICTURE_COUNT},
+  (_, index) => createPicture(index + 1)
+);
+
+getPictures();
